Roll Control
Cleaned up code that calculates roll angle of the forearm and uses that
to check for supination pronation.  The algorithm currently being used
to calculate roll is a corrected Aerospace rotation sequence.

The roll angle seems to change somewhat significantly with pitch angle
which is not good but this algorithm seemed better than the others.