Non-aerospace rotation sequence
With pitch and roll calculated using nonaerspace rotation sequence.
Roll angle does not flip signs when it reaches +- 90, however it is
desensitized with increasing pitch angle.  This means that if kids set
their thresholds in a horizontal position, they might not be able to
reach them with arm at other pitch angles.