Aerospace rotation sequence
Benefit is that the roll angle does not change significantly with pitch
angle.

Downside is that the roll angle changes signs at +- 90, which causes it
to switch between supination/pronation.  I added extra conditionals to
combat this, but now extreme pronation (past 90 degrees) reads as
neutral (I included a sketchy fix that seems to work)